requirements.txt
umpy
pandas
matplotlib
seaborn
scikit-learn
statsmodels
